Rename greenfield to Asset::Waveform and include as module.
