+++ app/models/asset.rb
@@ -3,8 +3,14 @@
 class Asset < ApplicationRecord
   include SoftDeletion
 
-  concerned_with :uploading, :radio, :statistics, :greenfield
+  require_dependency 'asset/waveform'
+  include Asset::Waveform
+
+
+  concerned_with :uploading, :radio, :statistics
+
   attribute :user_agent, :string
+  serialize :waveform, Array
 
   scope :published,       -> { where(private: false, is_spam: false) }
   scope :not_spam,        -> { where(is_spam: false) }
@@ -22,6 +28,7 @@ class Asset < ApplicationRecord
   belongs_to :user, counter_cache: true
   has_one :audio_feature, dependent: :destroy
   accepts_nested_attributes_for :audio_feature
+  has_one :greenfield_post, class_name: '::Greenfield::Post'
   has_many :tracks,    dependent: :destroy
   has_many :playlists, through: :tracks
   has_many :listens,   -> { order('listens.created_at DESC') }, dependent: :destroy
